requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-vision
google-cloud-storage
pypdf2
python-dotenv
pydantic
numpy
pandas
PyMuPDF